Move create-story task off registry-image resource
There seems to be a bug / missing feature for the registry-image
resource where it will fail to run if the image contains foreign layers.
I do not know what this means to us and why our image contains foreign
layers, but we will revert back to the original docker-image resource
for now.

There was an investigation that the BOSH Windows team did into this and
they have submitted a feature request to the upstream library which
handles the layers.

Links:
https://gist.github.com/genevieve/135474773f78c39a38717ca49083dda1
https://github.com/concourse/registry-image-resource/issues/33
https://github.com/google/go-containerregistry/issues/207