Fix newer usn-log not feeding into pipeline
* When the publish job clears the usn-log, this new
versioned usn-log was not feeding through the pipeline.
This is because "create-story-usn-" is the only job that
takes usn-log as an input. This meant that only when a
new CVE comes through will the empty usn-log flow.

* Fix is to have a dedicated job for "triggering" that takes
usn-log as an input. Conceptually, "usn-log" is similar to
"stemcell-trigger" in that many other jobs can produce a
new version, but there is a single job that is intended to
aggregate and start the pipeline

[#166618973]