Auto-bump BOSH blobstore CLIs
The blobstore CLIs are old. They're manually updated, which means
they're rarely, if ever, updated.

- AWS (s3cli/bosh-blobstore-s3)
- Google (gcscli/bosh-blobstore-gcs)
- DAV (davcli/bosh-blobstore-dav)

Also, the manner in which the BOSH blobstore CLIs are installed is
inconsistent--AWS & Google are installed in separate stages, whereas DAV
is installed as part of the BOSH Agent installation.

This commit addresses both these shortcomings by auto-bumping the BOSH blobstore
CLIs and making the installation consistent.

[#183017316] Bump old Golang dependencies in Opsman

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>