always use docker image build by 621 branch
fix git resource setup for docker image build. the job currently fails
with:

```
It doesn't appear that given Dockerfile: "bosh-linux-stemcell-builder/ci/docker/os-image-stemcell-builder/Dockerfile" is a file
```

that is because the added job for the dockerfile had copy paste errors
including the ytt templating for the stemcell-builder-repo branch
switching (to target the right branch for the stemcell line..).

while we have two branches for xenial, we only built the image with the
621 branch in the old pipeline.

this corrects the current pipeline to do the same.