Adds error handling for curl & jq calls
This commit adds some error handling for the curl and jq calls that
look for URLs to fetch USN data, and later fetch USN data from those
URLs.

It turns out that errors in Bash Process Substution pipelines will not
percolate up to the shell invoking the process substitution via the "-e"
or "-o pipefail" mechanisms. It turns out that you need to use the
'wait' builtin to have bash check the return value of the
most-recently-run process substitution.

By doing this, and relying on the apparent fact that the "-e" and "-o
pipefail" settings in the parent script appear to remain set in the
process substitution pipeline, we can detect errors in the process
substitution pipeline and abort processing before we erroneously
conclude that we have no USNs to deal with.

Signed-off-by: Brian Upton <brian.upton@broadcom.com>