Use correct metalink path for fetching USN data
`#{BRANCH}` is invalid as a metalink path when used in the Noble
pipeline. That is because the Noble branch does not have
a `bosh-stemcell/image-metalink/#ubuntu-noble` directory, while the
Jammy branch incorrectly has one where metalinks aren't updated. The
correct metalink file to use is
`bosh-stemcell/ubuntu-{jammy,noble}/ubuntu-{jammy,noble}.meta4`. The
correct branch is already checked out when this task is run.