add attempts when getting usn data
we have seen cases where the usn jobs will trigger, but when the resource tries
to get the oval data, it fails because it can't find the usn with that id (ex:
https://bosh.ci.cloudfoundry.org/teams/stemcell/pipelines/stemcells-ubuntu-jammy/jobs/log-low-medium-cves/builds/531).

From looking into this, it seems like the canonical endpoint we get the oval
data can be unreliable. If you rerun the job later, it will pass.

I'm hoping with a retry we'll have a better chance of getting the endpoint when
it is being cooperative.